NIFI-8497 Removed unnecessary model classes
